+++ source/main.cpp
@@ -28,9 +28,9 @@ int main(int Narguments, char* arguments[])
     // Platosim expects the filename of the configuration parameters.
     // Exit if this filename is not given.
 
-    if (Narguments < 3)
+    if ((Narguments < 3) || (Narguments > 5))
     {
-        cerr << "Usage: platosim <inputfile> <outputfile> [<logfile>]" << endl;
+        cerr << "Usage: platosim <inputfile> <outputfile> [<logfile>] [<logLevel>]" << endl;
         cerr << "       platosim -version" << endl;
         exit(EXIT_FAILURE);
     }
@@ -38,18 +38,39 @@ int main(int Narguments, char* arguments[])
     string inputFilename(arguments[1]);
     string outputFilename(arguments[2]);
     string logFilename = "log.txt";
-    if (Narguments == 4)
+    if (Narguments >= 4)
     {
         logFilename = arguments[3];
     }
 
+    int logLevel = 3;
+    if (Narguments == 5)
+    {
+        logLevel = atoi(arguments[4]);
+        if ((logLevel < 1) || (logLevel > 3)) 
+        {
+            cerr << "Error: logLevel should be either 1 (least verbose), 2, or 3 (most verbose)" << endl;
+            exit(EXIT_FAILURE);
+        }
+    }
+
     // Set up the log file
 
+    Log.addOutputStream(cerr, ERROR | WARNING);
+
     ofstream logFile(logFilename);
-    Log.addOutputStream(cerr,    WARNING | ERROR);
-    Log.addOutputStream(logFile, WARNING | ERROR | DEBUG | INFO);
+    switch (logLevel)
+    {
+        case 1: Log.addOutputStream(logFile, ERROR | WARNING);
+                break;
+        case 2: Log.addOutputStream(logFile, ERROR | WARNING | INFO); 
+                break;
+        case 3: Log.addOutputStream(logFile, ERROR | WARNING | INFO | DEBUG);
+                break;   
+    }
+    
     Log.info(string("PlatoSim ") + GIT_DESCRIBE);
-    Log.info("Main: Log file includes 'warning', 'error', 'debug', and 'info'");
+    Log.info("Main: Log file includes 'error', 'warning', 'info', and 'debug'");
 
 
     // Initialise the simulation, and loop over all exposures using run()